First pass: approval notifications to requester
I wrote a couple of classes for this which I now have a technical debt of
refactoring submit.php to use. :/

Also I should probably take out the collation fields and replace them
with function calls. :/

But, I'm going to go with it for now and write the Javascript which will
generate the collated fields, because I'm a terrible person.

I'm also thinking of reworking SQLActor use to work as a singleton
rather than instantiating it all over the place, but that is much less
necessary.