Removed all references to Consortia, haven't finished adding in Research Themes.
Also, have gotten a significant part of the way through putting user info and project request info into separate classes. And renaming things so that the emphasis on the request is on the project.

Should make it clearer.

I'm not sure yet whether to refactor a bunch of SQL stuff into the classes that use it.